+++ functions/fetch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @enschedule-fns/fetch
 
+## 0.0.3
+
+### Patch Changes
+
+- update package json and add worker-cli
+
 ## 0.0.2
 
 ### Patch Changes