+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @enschedule/cli
 
+## 1.1.3
+
+### Patch Changes
+
+- update package json and add worker-cli
+- Updated dependencies
+  - @enschedule/worker-api@1.1.3
+  - @enschedule/types@1.1.2
+
 ## 1.1.2
 
 ### Patch Changes